# pkg/trace: switch v0.5 msgpack deserialization to read from []byte