--- a/pkg/trace/pb/decoder_v05_test.go
+++ b/pkg/trace/pb/decoder_v05_test.go
@@ -6,45 +6,11 @@
 package pb
 
 import (
-	"bytes"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/tinylib/msgp/msgp"
 	vmsgp "github.com/vmihailenco/msgpack/v4"
 )
-
-func TestParseFloat64(t *testing.T) {
-	assert := assert.New(t)
-
-	data := []byte{
-		0x2a,             // 42
-		0xd1, 0xfb, 0x2e, // -1234
-		0xcd, 0x0a, 0x9b, // 2715
-		0xcb, 0x40, 0x09, 0x1e, 0xb8, 0x51, 0xeb, 0x85, 0x1f, // float64(3.14)
-	}
-
-	reader := msgp.NewReader(bytes.NewReader(data))
-
-	var f float64
-	var err error
-
-	f, err = parseFloat64(reader)
-	assert.NoError(err)
-	assert.Equal(42.0, f)
-
-	f, err = parseFloat64(reader)
-	assert.NoError(err)
-	assert.Equal(-1234.0, f)
-
-	f, err = parseFloat64(reader)
-	assert.NoError(err)
-	assert.Equal(2715.0, f)
-
-	f, err = parseFloat64(reader)
-	assert.NoError(err)
-	assert.Equal(3.14, f)
-}
 
 var data = [2]interface{}{
 	0: []string{
@@ -86,14 +52,12 @@ var data = [2]interface{}{
 	},
 }
 
-func TestDecodeMsgDictionary(t *testing.T) {
+func TestUnmarshalMsgDictionary(t *testing.T) {
 	b, err := vmsgp.Marshal(&data)
 	assert.NoError(t, err)
-	dc := NewMsgpReader(bytes.NewReader(b))
-	defer FreeMsgpReader(dc)
 
 	var traces Traces
-	if err := traces.DecodeMsgDictionary(dc); err != nil {
+	if err := traces.UnmarshalMsgDictionary(b); err != nil {
 		t.Fatal(err)
 	}
 	assert.EqualValues(t, traces[0][0], &Span{
@@ -118,17 +82,13 @@ func TestDecodeMsgDictionary(t *testing.T) {
 
 var benchOut Traces
 
-func BenchmarkDecodeMsgDictionary(b *testing.B) {
+func BenchmarkUnmarshalMsgDictionary(b *testing.B) {
 	bb, err := vmsgp.Marshal(&data)
 	assert.NoError(b, err)
-	r := bytes.NewReader(bb)
-	dc := NewMsgpReader(r)
-	defer FreeMsgpReader(dc)
 	b.ResetTimer()
 	b.ReportAllocs()
 	b.SetBytes(int64(len(bb)))
 	for i := 0; i < b.N; i++ {
-		r.Reset(bb)
-		assert.NoError(b, benchOut.DecodeMsgDictionary(dc))
+		assert.NoError(b, benchOut.UnmarshalMsgDictionary(bb))
 	}
 }