### What does this PR do?

This PR changes the msgpack deserialization for version 0.5 of the trace payload to read from a byte array instead of a byte reader.
 
It brings similar performance improvements to #6831 :
```
name                   old time/op    new time/op    delta
DecodeMsgDictionary-4    1.64µs ± 2%    1.13µs ± 5%  -31.33%  (p=0.008 n=5+5)

name                   old speed      new speed      delta
DecodeMsgDictionary-4   175MB/s ± 2%   255MB/s ± 5%  +45.69%  (p=0.008 n=5+5)

name                   old alloc/op   new alloc/op   delta
DecodeMsgDictionary-4      320B ± 0%      320B ± 0%     ~     (all equal)

name                   old allocs/op  new allocs/op  delta
DecodeMsgDictionary-4      12.0 ± 0%      11.0 ± 0%   -8.33%  (p=0.008 n=5+5)
```


### Motivation

What inspired you to submit this pull request?

### Additional Notes

Anything else we should know when reviewing?

### Describe your test plan

Write there any instructions and details you may have to test your PR.